file organization and responsive layout added
