+++ src/components/Landing.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import '../styles/Landing.css';
+
+const Landing = () => {
+    return ( 
+        <div>
+            <img 
+            src="https://cdn.jsdelivr.net/gh/MakeWeirdMusic/art@master/frank-zappa/frank-zappa.png"
+            id="zappaSVG" 
+            alt="zappa" />
+        </div>
+     );
+}
+ 
+export default Landing;
