+++ src/styles/App.css
@@ -8,6 +8,10 @@ body {
 
 .App {
   text-align: center;
+  display: grid;
+  grid-gap: 5px;
+  grid-template-columns: repeat(1, minmax(300px, 1fr));
+  grid-template-rows: repeat(1, 250px);
 }
 
 /* header */
@@ -23,7 +27,6 @@ body {
   width: 100%;
   background-color: white;
   box-shadow: 0 0 10px 10px white;
-  padding: 40px;
 }
 
 h1 {
@@ -33,6 +36,12 @@ h1 {
 #puncuation::before {
   content: " !";
   animation: askTell infinite 3s;
+}
+
+#puncuation {
+  /* font-family: 'Monofett', cursive; */
+  font-size: 60px;
+  cursor: pointer;
 }
 
 @keyframes askTell {
@@ -51,29 +60,6 @@ h1 {
   }
   100% {
     color: purple;
-  }
-}
-
-#play {
-  font-family: 'Monofett', cursive;
-  font-size: 60px;
-  cursor: pointer;
-}
-
-/* zappa svg */
-
-#zappaSVG {
-  width: 350px;
-  height: 350px;
-  transform: rotateY(180deg);
-  position: fixed;
-  right: -50%;
-  animation: slide 5s ease-in-out .5s forwards;
-}
-
-@keyframes slide {
-  100% {
-    right: 0;
   }
 }
